Branch b1_4 Merge minor changes from b_cray - asm/types.h build fix for catamount - include linux/quota.h only conditionally - use FPRIVATE in llite group locking code - use ptlrpcd and ptlrpcd-recov names for threads - add flock, group tests to b1_4 CVS - use OPENIBNAL instead of IBNAL for llvisualize - update lmc usage message

+++ lustre/liblustre/file.c
@@ -147,7 +147,7 @@ int llu_iop_open(struct pnode *pnode, int flags, mode_t mode)
         if (llu_is_root_inode(inode))
                 RETURN(0);
 
-        CDEBUG(D_VFSTRACE, "VFS Op:inode=%llu\n", st->st_ino);
+        CDEBUG(D_VFSTRACE, "VFS Op:inode=%llu\n", (long long)st->st_ino);
         LL_GET_INTENT(inode, it);
 
         if (!it->d.lustre.it_disposition) {
